+++ creme/feature_extraction/__init__.py
@@ -1,17 +1,18 @@
 """
 A module for extracting features from streaming data.
 """
-from .gb import GroupBy
-from .gb import TargetGroupBy
+from .agg import Agg
+from .agg import TargetAgg
 from .target_encoding import TargetEncoder
 from .vectorize import CountVectorizer
 from .vectorize import TFIDFVectorizer
 
 
 __all__ = [
+    'Agg',
     'CountVectorizer',
     'GroupBy',
+    'TargetAgg',
     'TargetEncoder',
-    'TargetGroupBy',
     'TFIDFVectorizer'
 ]